loadmem: load an ELF file from RAM
Add a new command, `loadmem`, to load an ELF object from
somewhere in memory.  This is useful for doing a quick test,
without the overhead of transferring a complete ramdisk.

Also fix a few typos in the README.  Thanks, Andy!

(Also remove the license checker for now; it does not respect
the MIT license on bits imported from Hypatia.)

Signed-off-by: Dan Cross <cross@oxidecomputer.com>