zoxboot: an alias for the most common boot case
Add the skeleton of an alias system, and provide a default for
the most common case of booting a machine with a ramdisk sent
via ZMODEM.  With this change, the single command `zoxboot` will
automatically receive a compressed ramdisk (either UFS or a cpio
miniroot), expand it, mount it, load our kernel nucleus from it,
and call the kernel's entry point.

Tested on Garnet.

Signed-off-by: Dan Cross <cross@oxidecomputer.com>